+++ crates/bitwarden/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Support for fetching multiple secrets by ids (#150)
+
 ## [0.3.0] - 2023-07-26
 
 ### Deprecated