+++ crates/bitwarden/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `auth::request::AccessTokenLoginRequest` moved to `auth::login::AccessTokenLoginRequest` (#178)
 
+- Support for fetching multiple secrets by ids (#150)
+
 ## [0.3.0] - 2023-07-26
 
 ### Deprecated